Remove nightly backup of Asset Manager assets from NFS to S3
Once we're happy that cross-site replication from the
govuk-assets-production S3 bucket to the govuk-assets-backup-production
S3 bucket is working OK, we should be able to remove this
asset-manager-s3 job which does a nightly backup of Asset Manager assets
from NFS to the govuk-offsite-backups-production S3 bucket.

Unfortunately I haven't been able to test that specifying `ensure:
'absent'` removes the relevant resources, because this job is only
declared in the production environment. However, reading the code it
looks as if it should work.

Note that the assets-whitehall-s3 job should stay in place until all
existing Whitehall assets have been migrated to Asset Manager.