Pass ensure param from offsite backup job to Icinga check
I want to be able to disable one of the offsite backup jobs. I'm
planning to do this by adding `ensure: 'absent'` to the config for the
relevant job in the `backup::assets::jobs` hash in
`hieradata/production.yaml`. However, prior to this change, doing the
above would *not* have removed the corresponding Icinga check and it
would eventually have triffered an alert.

By removing the `if` condition and passing the `ensure` parameter
through to the Icinga check, the documentation [1] suggests that the
latter will be removed if I set `ensure` to 'absent' as described above.

Note that the default behaviour (where `ensure` is set to 'present')
should remain unchanged.

[1]:
https://github.com/alphagov/govuk-puppet/blob/a3a27155782f779b1e79c7f268ca76f1c9e26ec8/modules/icinga/manifests/passive_check.pp#L19-L20